fix lesson 16

+++ resources/views/article/index.blade.php
@@ -9,7 +9,7 @@
         <h2><a href="articles/{{$article->id}}">{{$article->name}}</a></h2>
         <small>
             <a href="articles/{{$article->id}}/edit">Редактирование</a> |
-            <a href="articles/{{$article->id}}" data-method="delete" rel="nofollow">Удалить</a>
+            <a href="articles/{{$article->id}}" data-confirm="Вы уверены?" data-method="delete" rel="nofollow">Удалить</a>
         </small>
 
         <div>{{Str::limit($article->body, 200)}}</div>
